Updating Windsor to 3.0.0 fixes #1
Updated Windsor and used the new `Scoped` lifestyle to register modules.
A customized `NancyWindsorContextFactory` begins the scope and adds it
to the NancyContext. Everything is scoped to the call context.

An interceptor is used to intercept the call to
INancyEngine.HandleRequest() and wraps this call in the Scope. When this
call ends the scope is disposed of.

In the cases where GetModules() and GetModulesByKey() are called outside
of the context of a HandleRequest call we create a temporary scope to
obtain them. This scope is disposed of immediately after obtaining the
references.